refactor: added additional error handling within the backend 911 controllers and fixing api routes
